requirements.txt
pandas
numpy
prometheus-client
requests
psycopg2-binary
